requirements.txt
# ============================================================================
# LLM Microservice Python Dependencies
# ============================================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation & Settings
pydantic>=2.7.0  # Upgraded for elevenlabs compatibility (requires pydantic-core>=2.18.2)
pydantic-settings>=2.1.0
email-validator==2.1.0

# LLM Provider SDKs
anthropic>=0.40.0
openai==1.3.7
groq==0.4.1
google-generativeai==0.3.2
mistralai==0.0.12
cohere==4.37
runwayml>=1.0.0
fal-client>=0.4.0
elevenlabs>=2.21.0
voyageai>=0.2.0
assemblyai>=0.35.1
deepgram-sdk>=3.0.0
google-api-python-client>=2.100.0  # Perspective API (Google Discovery API)
google-auth>=2.23.0
httpx==0.25.2

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Cache & Storage
redis==5.0.1
minio==7.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Token Counting
tiktoken==0.5.2

# Monitoring & Metrics
prometheus-client==0.19.0

# Logging
structlog==23.2.0

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
